Library/SctLib: Removed dependency on UefiServicesTableLib in Shell.c
The mix of INF version in the SCT build causes some inconsistent views
in the depedencies.

ARM Toolchain (RVCT) was failing to build because of undefined
gImageHandle which was defined by UefiBootServicesTableLib but
not included by the EDK2 Build system into SctLib when invoked
by a module using an older version of the INF file.

Change-Id: Ie9f0ee877027fa650739d92cc26907279c3d6f94