Add script to automatically generate test file structure.
When new_test.py is run with no options, it asks the user a series
of questions about the test they wish to create and then creates the test
file structure, files, shell functions, and entries in build system
configuration files.

When run with the option -g or --guid, it just generates a GUID in a form
that matches the format used in code.

Signed-off-by: John Powell <john.powell@arm.com>